options for objectives

+++ code/src/include/utils.hpp
@@ -4,7 +4,7 @@
 #ifndef _PRIMER_UTILS_HPP
 #define _PRIMER_UTILS_HPP
 
-#define INFTY static_cast<size_t>(numeric_limits<int>::max())
+#define INFTY static_cast<size_t>(numeric_limits<size_t>::max())
 
 // #define DEBUG_MODE
